SS-164: Making changes to the create shift window layout based on new data-model and UX design
